ticv02.l: separate etiquettes from data processing
Following the previous commit and to avoid a similar issue in standard
mode, apply the same recipe:

Use an exclusive start condition to process data.
Logic: when we encounter the first SEP in INITIAL condition, everything
that comes after is DATA and not etiquette until we start a new dataset.
Only change states at delimiters and use condition catchalls for them to
ensure resync in case of data errors.
This makes initial frame sync a little more error-verbose.