Add build recipes for ESP8266/ESP32 FreeRTOS IDF
This commit adds recipes for a quick implementation of tic2json on
Espressif FreeRTOS IDF SDK. It builds identically for ESP8266 and ESP32
(tested only on ESP8266).

The included main_app() stub sets up a serial conversion with raw TIC
on input (RX) and JSON on output (TX).

The build recipe is hackish, but it get things done.